refactor location files
