--- src/pages/main-page/constants.ts
@@ -1,3 +1,0 @@
-export const DEFAULT_MAP_CENTER: [number, number] = [
-  52.3909553943508, 4.85309666406198,
-];
